Updated gridPars
In older versions, we have plotted M1 (ref) in gray and M2 (target) in
black. It was switched when we introduced gridPars but I have reverted
it back now.

Also reordered the parameters so all the reference ones come first,
then the target ones, the the grid ones, then the text ones.